Fixed macos bug in main.c
Game would crash because infoPanel wasn't null when checking for user input on the menu. For some reason macos was setting the variable to 0x3f, but making the var non-static fixed the problem. I shouldn't use static anyway, but I don't know why it was set to 0x3f, because it was still NULL at the start of the program...