Fixed a weird Windows bug
It no longer takes 45 minutes to load the game, and the colors are correct. Turns out changing colors on the windows terminal doesn't seem to work properly, but can_change_color() returns true. I fixed it with a really bad hack that defines can_change_color() as false, so that the program doesn't try to change color